rename 'queries' to 'from_table' and move under 'linters' module (#2568)
## Changes
`queries.py` is improperly named and should reside in the `linters`
module
This PR makes these changes.

### Linked issues
None

### Functionality
None

### Tests
- [x] checked unit tests
