+++ src/databricks/labs/ucx/source_code/linters/from_table.py
